service/pyproject.toml
[project]
name = "ddd-fastapi-service"
version = "0.1.0"
description = "Production-ready FastAPI backend with DDD/Clean Architecture and LangGraph"
authors = [
  { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.11"

[project.scripts]
app = "app.main:run"

[tool.uv]
# uv will create and manage a local virtual environment
managed = true

[dependency-groups]
# runtime dependencies
default = [
  "fastapi>=0.116.2",
  "uvicorn[standard]>=0.35.0",
  "pydantic-settings>=2.10.1",
  "python-dotenv>=1.1.1",
  "httpx>=0.28.1",
  "langgraph>=0.6.7",
  "langchain-core>=0.3.76",
  "typing-extensions>=4.15.0"
]

# development-only dependencies
dev = [
  "pytest>=8.4.2",
  "anyio>=4.10.0",
  "ruff>=0.13.0",
  "mypy>=1.18.1",
  "pytest-asyncio>=1.2.0",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E203", "E266", "E501", "SIM115"]
exclude = [".venv", "venv", "build", "dist"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["src/app/tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.sources]
app = "src/app"
